Tag button features with future release
Because of a breaking change to the button styles the current master
build fails as browserstack tests run against the current github-pages
URL which is not reflective of what's in master.

For now I'm using the `@future_release` tag to mark those tests as
pending release. Longer term it would be great to get the release
process to a point where browserstack tests can be run against the same
thing as the build.