Fix flakiness of button active state test
Because we are using a custom onReady script for the button active state
visual regression tests the `prefers-reduced-motion` setting was not
being set. Without it the test can by flaky as the test may be run
mid-css-transition. This PR adds that setting and regenerates the reference image to include that change.