Remove callout.js
Now that we handle determining what heading level to use at render time
for callouts we don't need code which attempted to do this client-side.

I thought about using the same deprecated pattern we use in tables.js
but realised that no projects have ever called this module so it's safe
to just delete it.