Simplify targeted content feature tests
Makes two main changes:

1. Removes welsh language specs in favour of rspec equivalents
2. Runs all browser scenarios against the "anchored" example.

The removal of welsh language specs is done on the basis that longer
term we aim to keep the welsh language behaviour out of storybook. In
the context of the design system translations are specifically bundled
for use with Rails only, so it makes sense to keep the tests within the
engine specs.

The merging of the "default" and "anchored" rules is down to the fact
that "anchored" is not a separate variant, it's an example within
storybook that shows the regular targeted content in the context of a
full page. This is a much more realistic test scenario so we can stand
to run all the existing scenarios against this example.