Sync ruby versions
In https://github.com/citizensadvice/design-system/pull/1187 ruby was
updated to `2.7.3` but only in some places.

This builds on that and makes sure we use the same version everywhere.

In the process I've merged the engine workflow in with the main one as
they don't need to be separate workflow files and it allows us to share
the version more easily.