Header: Update styles to rely on cascade
In order to have maximum flexibility when handling account links I've
adjusted the styles to rely more on the cascade and style anchor and
button elements directly within the header links section without relying
on people needing to remember the right classes when rendering a custom
form. They can provide a <button> and we'll take care of the styling.