Remove title property from callout
We have a `title` property which confusingly maps to an aria-label.
Based on the real world usage of this there's no scenario wherre we
provide this that isn't otherwise covered by the content or headings
within the callout itself. Fairly sure this is a bug and we shouldn't be
setting an aria-label here.