Allow demo-app job to run earlier
Extracted from https://github.com/citizensadvice/design-system/pull/1841

This is to a) allow us to run the demo app job earlier as it will be doing more work b) avoid running lint checks for very single version of ruby/rails. We only care about rspec tests for different versions. We'll need to add this job to the required checks after merging this.