Add consistent dev script for demo and docs site
Switching between the demo and docs site it would be useful to have a
consistent start up script. A common Rails convention for this seems to
be to have a bin/dev script.

**Bridgetown**

Adds a basic bin/dev script which delegates to bin/bridgetown start

**Demo app**

Uses foreman along with a Procfile to run the server and
webpack-dev-server in a single command

Pinched from https://github.com/rails/cssbundling-rails

---

The upshot is that for each sub-project you can consistently run
`bin/setup` to set up the project and `bin/dev` to start a dev server.