Update documentation around testing
Not compiling the packs for the `test` environment was a real gotya for
me coming to this new.

The `check` rake does not appear to exist so I removed the reference.

I think there is a case to run the webpack dev server in the test
evironment here, but that is a much more controversial change so I am
going with adding the docs to help others.