Updates after running a fresh setup
With the recent experience with external contributions and with the
addition of consistent dev scripts I thought I'd try a fresh clone of
the project to check for any other pain points.

Adds a couple of extra steps to the top-level setup script:

1. Runs a build after the initial install step
2. Adds the docs set up script to the top-level one

This should mean that you can run `bin/setup` at the top-level followed
by `bin/dev` within either the demo app directory or docs site directory
to get a working install.