Compile assets in demo app test environment
This should improve the new setup workflow a bit as it stops cypress
tests from relying on a specific test asset build.

We don't really care about different environments for the demo app so we
can stand to have more parity between them. Allows the test environment
to automatically compile assets if it needs to and bundles them in the
same directory, no need for a separate packs-test directory.

Simplifies some of the documentation now there should be fewer gotchas.