Move button backstop scenarios down.
We occasionally see an error with the button backstop scenarios where
they error with a ActionNotFound error or otherwise timeout. Curious to
test if this is specific to the button scenarios or if it's because they
come first. Test this assumption by moving the button scenarios down a
bit (greater than the `asyncCaptureLimit`)