Use capybara page method in tests
As of the latest version of view_component rendered_component was
renamed to rendered_content and subsequently marked as deprecated.

Apparently the intent was for this to be a private method and for tests
to prefer using the capybara `page` method.

Most of our tests already do this, so replace all remaining uses of
`rendered_content` with `page`.