Header/Footer: Fix ambiguous logo label
The logo has a couple of accessibility issues flagged by cypress-html-validate

1. We use title instead of aria-label
2. We link the logo in both the header and the footer with typically the
   same title or label

The former is easy to fix, use aria-label instead. For the second point,
the simplest thing to do is to not link the logo in the footer at all as
it's decorative and repeating the role of the header logo.