Upgrade to stylelint v14
Restructures stylelint dependencies and to use stylelint-config-standard-scss
in favour of manually configuring scss syntax as recommended in the docs.
This automatically handles the new customSyntax requirements from the new
stylelint version and reduces the number of direct packages we manage.

Updates rules to reflect the new config and any new rules in stylelint v14.
Removes any rules that are defined by newer versions of the standard config.