Allow opting-out of external link styles
A recent accessibility report flagged that our external link icon can be
interpreted as a new tab icon causing confusion.

We have seen similar reports in testing sessions for the new intranet.

This mirrors the issues mentioned in https://designnotes.blog.gov.uk/2016/11/28/removing-the-external-link-icon-from-gov-uk/

We also know that we have an outstanding accessibility issue with our
icon fonts which affects this icon too https://github.com/citizensadvice/design-system/issues/1902

To allow us to test removing the external link icon, add a new config
option to opt-out of the external link styles:
`$cads-enable-external-link-icon`

This is enabled by default to respect the current defaults.