Demonstrate output buffer bug
As shown in https://github.com/citizensadvice/design-system/pull/2179
there is an issue rendering components in bridgetown where blocks are
used.

This causes an output buffer issue where the contents of the block is
also rendered on the page.

This only seems to happen with haml templates not erb templates. It's
possibly a bridgetown bug but more likely down to the fact we are having
to approximate haml support in our builder by having to patch in a new
template handler.