Fix output buffer issue by switching to ERB
This change demonstrates how switching to ERB for internal component
templates bypasses the output buffer issues seen in
https://github.com/citizensadvice/design-system/pull/2207 when using
design system components within bridgetown.

It's possible to workaround some of the issues by adjusting the haml
templates with delicate use of `capture_haml` or by adjusting other
parts of the templates as shown in
https://github.com/citizensadvice/design-system/pull/2179 but this feels
fraught.

By switching to ERB for internal component templates we'd be able to
avoid the need for any custom templating logic in the builder as ERB
is a core part of ruby so has a higher likelihood of working without
caveats and be officially supported.

Because view component exists to wrap the underlying templates in an
interface we don't need to change all the templates in one go—only the
affected ones—and it doesn't affect our ability to use haml in the main
apps.