Run Backstop tests in test mode again
In https://github.com/citizensadvice/design-system/pull/2210 we made
some changes to improve the reliability of the Backstop tests.

One of those changes was to run the app in production mode. It turns out
that this:

a) isn't necessary, the real fix was waiting on a specific
component URL before running the tests and
b) whilst builds work fine for normal pull requests they
fail for dependabot changes because it doesn't have access to the secret used for SECRET_KEY_BASE.

As it's not actually necessary to run in production mode let's swap back
to test mode but with eager-loading enabled in CI which should do the
job without breaking on dependabot changes.