Navigation: Add background to colour language
Extracted from https://github.com/citizensadvice/design-system/pull/2251

Currently the colour of the navigation is controlled by the
primary brand colour. This is a problem if you need to control the
colour of the navigation without changing the brand colour as is the
case in https://github.com/citizensadvice/design-system/pull/2251

Update the navigation styles to use a dedicated variable for the
navigation background colour.