Introduce the concept of scoped prose styles
Often we use the cads-prose class for content where we also embed lots
of rich components inside. By default styles within cads-prose cascade
down which can subtly affect any components inside. We often work around
this in our styles by resetting styles if we are inside a cads-prose
element.

I'm proposing that we introduce a new cads-prose-scoped class which uses
a child combinator selector to only apply typographic styles to
immediate children. Any components within a scoped prose block would not
be effected as they would start a new context.

I think the way our standard cads-prose class behaves is a sensible
default but it is useful to have an equivalent class that is more
targeted for cases where we need a bit more control.

---

Notably using the child combinator in combination with an extend logs a
warning that extending this kind of selector will be a warning Dart Sass
2.0.0

Using a mixin here has no major impact as our minifier already
dedupes selectors (doing what extend does but automatically).