Fix shadowing issue with SectionLinks component
Currently the SectionLinks component accepts a single section_links
slot plus optional content for any custom content (e.g adviser only).

This causes an unexpected issue in a bridgetown context where the
content variable gets shadowed by the wider rendering context.

This points to an API design issue with the SectionLinks component.
Typically we only use the default content block when either a) the
content is always required b) the content is optional and there are no
other slots. Having a slot but **optional** content creates are weird
situation where we are passing content block but typically no content.

We should follow the pattern we use for other slots based components
like the header and prefer slots for everything here.

In this case this means providing a `custom_content` slot to configure
any custom content rather than relying on content being passed at the
top-level.

To avoid this being a breaking change I've added this a new slot and
added a deprecation warning for the old content block. Wrapping it in an
extra `respond_to?(:to_str)` check avoids the ghosting effect in
bridgetown which is not affected using the new slot so can be removed in
a later major version.

Converts the template to ERB in the process of this change.