Publish docs on main build
Currently our docs build workflow only publishes drafts and on release
versions.

Whilst we are still working on the foundations of thenew docs website it
would be nice to have a continuously deployed version of the latest
changes. Added a new workflow that publishes on main builds which we can
remove later if we want to stick to release-only builds.