Error summary: Make autofocus optional
Updates the error summary component to allow opting out of the autofocus
behaviour. This is really only useful for the docs website where
autofocusing the examples is undesirable, as such I've left it
undocumented in the options table.

Expands the tests to cover this new scenario plus to test the untested
heading_level option and use capybara matchers.

Converts the template to ERB in the process.