Upgrade size-limit dependencies
Noticed that size-limit has stopped working for our JS files (always
reporting 0 B).

Uses this as an excuse to upgrade size-limit to the latest version. The
latest version uses esbuild now for size-limit/preset-small-lib which
doesn't make sense for our css files. All we really care about is a
file-size check so switches the setup to use the basic file plugin and
updates the glob for our JS files to include everything, not just
top-level files.