Add a stylelint rule for cads- prefixed classes
We currently use plugin/selector-bem-pattern which is run against
component styles but this a) requires and explicit comment to be added
to the components and b) only really applies for components that use
strict BEM.

Adds a more general selector-class-pattern which checks that any class
selector uses the cads- prefix (with a special case for .no-js).

This caught some known modifier classes for icons and the greedy nav
(although maybe this should still be prefixed with cads in future).

It also caught one stray .paragraph-s class in the notice banner
component which had no accompanying markup or base styles so can be
removed.