Add implicit rules for BEM stylelint plugin
Previously plugin/selector-bem-pattern required an explicit comment to
check for BEM naming conventions for component and utility styles.

Adds implicit config rules for any object, component, and utility styles
to force us to always define these; even if it's using the weak setting.

Requires a small change to the visibility classes to avoid generating
class names so the linter can pick up on them.

As we only generate these classes for medium and large breakpoints this
is only a small amount of duplication.