Fix ruby 3 syntax issues in docs and demo app
See https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/

Ruby 3 separates how positional and keyword arguments are handled.
Passing in arguments to the single checkbox as a hash relied on a quirk
of ruby 2 where hashes were converted to positional arguments.

Now we either need to pass them explicitly or use the ** double splat
operator. It's simplest to use the named parameters directly here.