Fix label being assigned as checkbox attribute
When running against ruby 3 and passing options to a checkbox via
c.checkbox(**my_checkbox_arguments) the previous **additional_attributes
keyword args would include the named label which we don't want.

This is down to the `super` call in the initialise method.
Calls to super in subclasses of ViewComponent::Base aren't needed
by default and can change the behaviour of **kwargs so we exclude
component classes from this rule.