Clean up component specs
Does a bit of clean up on the component specs, mainly to:

1. Convert remaining specs to use capybara matchers
2. Fix rubocop warnings in locale specific tests
3. Clarify any scenarios where the previous specs were misleading

Includes a few notable updates:

**Checkbox group specs**

Splits out the optional arguments scenarios into multiple blocks. In the
process clarifies that the layout and size options don't apply to
checkbox groups, only radio groups.

**Radio group specs**

Expands specs to cover the same scenarios as checkbox groups. Whilst
they inherit a base class they are ultimately separate interfaces so we
should test both fully. Expands specs for layout and size options which
are unique to radio groups.

**Text input and textarea specs**

Expands text input and textarea specs to cover the same scenarios as the
base input class. Whilst they inherit a base class they are ultimately
separate interfaces so we should test both fully. This ultimately allows
us more flexibility if we need to decouple their interfaces in future.

**Targeted content specs**

Fixes an issue where the no content scenario tested against the callout
component by mistake.

**Table specs**

Remove rspec-snapshot. This is the only set of specs that uses it so
we are better off replacing this with more explicit matchers.