On this page: Fix css columns overflow bug
== Background ==

The on this page component has three main variants:

- A basic layout (fewer than four items) which is a simple list of links
- A multi-column layout (four or more items) which uses css columns
- An expandable layout when `show_nested_links` is set

== Issue ==

In newer versions of Chrome when viewing the expandable version of the
component at a non-default zoom level it is possible for the links to
overflow when expanding a list of child links.

After some investigation, the cause of the issue was down to the way we
were using css column styles for this component variant.

The way we had the CSS set up for the expandable variant to unset the
column layout by using `columns: 1`. This along with having
`break-inside: avoid` set to avoid columns being split when the wrapped
caused the underlying issue. CSS column calculation is notoriously
buggy; thankfully in this case we can avoid it entirely.

== Changes ===

Because the multi-column layout is really an exception, rather than the
rule, we can adjust our styles to scope and multi-column styles to only
the variant that needs them i.e. rather than setting `columns: 1` on our
expandable variant simply use default list styles.

This avoids any layout calculation bugs with css columns, especially as
we don't need it.