Remove rails main from test matrix
There's a downstream issue with the edge version of rails that is
causing our builds to fail.

Given that for our use case we have a relatively small number of known
consumers and know the exact rails versions they all run we probably
don't need to be testing against the bleeding-edge version of rails.

Removing it should fix the current build failures.