Use with slot format in specs
This is extracted from https://github.com/citizensadvice/design-system/pull/2758

In recent versions of viewcomponent a new format was introduced for
using slots, using a `with_` prefix.

See: https://viewcomponent.org/adrs/0004-slots-separate-getter-setter.html

This is optional is version 2 but required in version 3 which has now
been released. We can prepare for this upgrade ahead of time by
replacing any internal slot references with the new format, starting
with specs.