Remove generated example files
These previously existed as a bridge between the view component previews
and storybook. They are no longer connected to anything but we kept them
around as we anticipated that they might act as a useful reference.

In practice these have not been updated since we removed storybook and
the documentation website and demo app provide a better implementation
reference than relying on these static example files, which have begun
to drift having not been updated in several months.