Simplify typescript set up and package exports
Currently we have a couple of papercuts with our packaged javascript
modules

1. The method of importing and initialising them is inconsistent
2. The typescript configuration needed is relatively complex for a small
   number of modules

This commit attempts to rationalise that so that we have a consistent
pattern for importing and initialising all modules and as a side-benefit
a simpler typescript configuration.

- Defines a top-level tsconfig.json that works for all modules and
  removes any per-module configuration needed
- Adds consistent index.ts files to every module with a flexible set of
  default and named exports
- Adds a wrapper method to greedy nav to allow it to be initialised in
  the same way as all other modules

The goal is to support the existing import paths whilst enabling a
consistent preferred method using named exports:

```
import { initMyModule } from
"@citizensadvice/design-system/lib/my-module"
```

I've also included a `lib/index.ts` file so you can import from a single
entrypoint:

```
import { initModuleA, initModuleB } from
"@citizensadvice/design-system/lib"
```

Ideally we'd support importing directly from
`@citizensadvice/design-system` but that needs further restructuring to
make this work so something for a future release.