Refactor greedy navigation
Summary
=======

Introduces a new version of the greedy navigation component with
greatly simplified code and improved behaviour.

Motivation for change
=====================

The greedy navigation component is one of our oldest components and was
adapted from an external component. It's long been a pain point of the
application as it is hard for the team to modify leading to patch work
amends and redundant behaviour being maintained.

The main motivation for change is that the component uses the icon font
for the toggle button and we'd like to get rid of this in favour of SVG
icons for all components.

This requires change. The component is extremely hard to change so we
should make the change easy by restructuring the component to only
contain the behaviour we need and to match the rest of our components.

The old component also has a completely different API to the rest of our
components so this is a good opportunity to make that consistent.

Configuration free
==================

The new component is configuration free. The old component provided a
number of configuration options many of which were partially implemented
or redundant.

The exception being label text which is now translated as part of the
Rails template.

The new component is initialised like all our other components.

```js
import { initNavigation } from '@citizensadvice/design-system/lib';
initNavigation();
```

Improvements
============

Restructuring the component enables a number of improvements:

1. Switches the icon to an SVG version (the original motivation for
   change)
2. Improves keyboard and mouse support (tabbing and clicking in and out
   of the menu now behaves more consistently)
3. Minimises flash of unstyled content (fix ported from the public
   website)
3. Bundles welsh translations for UI text. Previously this had to be
   provided as part of the component initialisation (translations taken
   taken from the public website).
4. Fixes mistranslation of aria-labels (in production when translated
   they are part welsh part english)
5. Removed significant amounts of redundant behaviour that was brought
   over from the module we modified.

Backwards compatibility
=======================

To make it easier to migrate to the new component I've kept the old
entrypoint around which wraps the new implementation. This allows us to
ship the new version with everything need as part of the library and we
can remove the old entrypoint in a subsequent version.