fix: use `import` instead of `use` for overridable grid settings
This makes overriding the grid settings possible in products that use the design system.  `@use` requires you to explicitly override variables when you import a module, rather than respecting what the global state of the `@use`d file is.  This changes ensures that the global values of the grid are respected at all layers of the ITCSS triangle, allowing products to override the variables to make their own grids.