Patch bridgetown compatibility with view_component 3.9.0+
ViewComponent 3.9+ introduced a change which depends on there being an active Rails
request or controller in order to conditionally escape HTML. This causes an issue
when run in bridgetown as there's no request and we're always rendering static html.
Monkey patch the relevant method to remove the request format check.

Fixes #3145