Simplify setup scripts
Originally, sub-projects defined setup scripts which would not work if
you hadn't first run the top-level one…which then ran the sub-project ones.

This is pretty confusing and something we want to avoid.

Consolidate on a single top level `./setup` script and remove
sub-project ones to remove ambiguity.

Also update the Rakefiles to provide a more consistent
interface when running different checks.