Docs website: Upgrade view_component with patch
Upgrades view_component in the docs website with an additional
compatibility patch.

In version 3.15+ the docs build fails with the following error:

```sh
undefined method `controller' for an instance of Bridgetown::ERBView
```

This is similar to the error that was patched in #3362 but likely
resurfacing because we applied the narrowest possible monkey patch
originally.

Adjusts our patch for Bridgetown compatibility with view_component.
Substituting in a controller covers what the library expects to be present.

Using `Bridgetown::ViewComponentHelpers` also conflicts with the
presence of a `request` object. It turns out we don't need this for our
use case so remove the include rather than attempting to patch around
it.