Refine summary list styles
Restructures the summary list styles to simplify the padding around
elements and avoid the need to override as much by using the lower level
typographic scale mixin rather than an extend.

Defines a margin for the top-level component which is in keeping with
our spacing scale. The inherited value was the default for definition
lists which was tight when combined with other elements.