Pin concurrent-ruby version for tests
Workaround for
https://github.com/ruby-concurrency/concurrent-ruby/issues/1077 which
is currently causing our Rails 7.0.x CI tests to fail.

This change shouldn't affect the published gem as this only controls
what we install locally or in CI and consumers control their own
application gem versions.

Can be removed when either a Rails 7.0.x patch is released
or we remove support for Rails 7.0.x whichever is sooner
(likely the latter as it is near EOL)