Refactor on this page to use SVG icons
This change removes the use of icon fonts from the on this page
component. To enable this the component has been refactored to inline
the toggle logic rather than sharing it between two components to make
per-component changes simpler to manage.

Includes an additional cypress tests to make sure we have test coverage
as some is lost by inlining the shared code.