feat: allow breakpoints to be overriden by products
This makes overriding the breakpoint settings possible in products that use the design system.  `@use` requires you to explicitly override variables when you import a module, rather than respecting what the global state of the `@use`d file is.  This changes ensures that the global values of the breakpoints are respected at all layers of the ITCSS triangle, allowing products to override the breakpoint variables as they wish.