Set engine version
We currently fix the engine version to 0.1.0 which prevents us from
using things like the deprecator to log version warnings and is at best
just a little confusing.

Set the engine version to match the package version and add a spec to
require them to match when updating the version.