Migrate rails form builder to design system engine
Background
==========

We've historically maintained a separate rails-form-builder gem for
using the design system components as a standard Rails form builder.

This was done to allow iterating on the rails form builder concept
independently of design system releases.

Whilst this current set up was faster to iterate on the internal API
changes to components themselves required multiple steps before any
new updates could be implemented leading to drift as can be seen through
the change to introduce a page_heading option to form fields (#3495)
which was never implemented in the rails form builder.

As the interface for the library has now stabilised and we've reached
the point where we are using the library in multiple public production
systems it's a good time to bring the form builder into the main design
system engine.

Changes
=======

Ports all the code from citizensadvice/rails-form-builder to the design
system engine. Includes some updates to the demo application to use the
form builder along with a basic Cypress test.

A companion ADR is also included.