Add scaffolding for character counter
Background
==========

We currently have a basic character counter in use on the referrals
application. It's general enough and we have an open issue (#1691) that
suggests it's worth making this available through the design system
itself.

Changes
=======

Adds the scaffolding for a character count component supported by
textareas. Takes care to follow the same accessibility patterns as the
GOV.UK equivalent which has recently undergone some major accessibility
improvements[^1].

This version also differs from the current version in our products in that it
doesn't use `maxlength` as the limit. This is deliberate to avoid
data-loss when pasting text[^2].

As a result this version has the following additional states:

1. A no-js fallback state
2. A error state when exceeding the character limit
3. A screen-reader only version of the character count which updates at
   a slower rate.

Caveats
=======

The current implementation doesn't have any tests (neither rspec updates
or Cypress tests) nor documentation.

Using this as a starting point to validate against the current implementation.

[^1]: https://dav-idc.com/making-a-character-count-component-more-accessible/
[^2]: https://adamsilver.io/blog/dont-use-the-maxlength-attribute-to-stop-users-from-exceeding-the-limit/