Add Rails helpers for working with breadcrumbs
We've established a pattern in our application for working with
breadcrumbs from within controllers and views. This pattern is even
included in the demo application in this repo.

Extract the pattern into the recently added
`CitizensAdviceComponents::Helpers` concern.

Updates the using with rails guide and the component guides to document
the helper behaviour as well as covering some of the recently added page
classes.

Also updates the breadcrumb component itself to accept the array of
links as the first argument, rather than using a positional argument,
for a cleaner interface. Supports the old interface via a deprecation
warning.