bugfix: when header links are collapsed into the nav menu, always show the nav menu when the header links are hidden
this fixes a bug (NAT-1407) on the Intranet, where there are not that many header links so it does not collapse down
to the dropdown menu on mobile, but because the header links are hidden and only behind the More dropdown, it becomes
impossible to then access those header links. This changes the behaviour so that the More button is always shown if
the header links exist but are not visible.