Support ViewComponent 4
ViewComponent has published a release candidate for version 4. This
version targets a Long-Term-Support model meaning that breaking changes
will be less frequent going forwards.

To help prepare for this release introduce Appraisal to help us simplify
how we manage tests across supported versions.

Includes an ADR for the introduction of Appraisal along with a new
maintenance policy guide for contributors and updates to the testing
guide.